latest push

+++ src/app/faculty-add-notes/faculty-add-notes.component.ts
@@ -5,7 +5,7 @@ import { observable, Observable, Subscriber } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { NotesInterface } from '../notes-interface';
 import { ActivatedRoute, Router } from '@angular/router';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, ConfirmEventType, MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-faculty-add-notes',
@@ -21,7 +21,7 @@ export class FacultyAddNotesComponent implements OnInit {
   notes: {};
   notesData: NotesInterface;
   notesid = 0;
-  constructor(private service: DataService, private route: ActivatedRoute,private rut:Router, private messageService: MessageService) { }
+  constructor(private service: DataService, private route: ActivatedRoute,private confirmationService: ConfirmationService, private rut: Router, private messageService: MessageService) { }
 
   async ngOnInit() {
     this.userid = parseInt(sessionStorage.getItem("userid"));
@@ -41,7 +41,7 @@ export class FacultyAddNotesComponent implements OnInit {
           notesid: new FormControl(this.notesData.notesid, Validators.required),
           userid: new FormControl(this.notesData.userid, Validators.required),
           courseid: new FormControl(this.notesData.courseid, Validators.required),
-          path: new FormControl(this.notesData.path, Validators.required)
+          path: new FormControl(this.notesData.path, [Validators.required,Validators.minLength(2)])
         })
       })
 
@@ -50,7 +50,7 @@ export class FacultyAddNotesComponent implements OnInit {
       this.myform = new FormGroup({
         userid: new FormControl(this.userid, Validators.required),
         courseid: new FormControl("", Validators.required),
-        path: new FormControl("", Validators.required)
+        path: new FormControl("", [Validators.required,Validators.minLength(2)])
       })
     }
 
@@ -58,45 +58,58 @@ export class FacultyAddNotesComponent implements OnInit {
   }
 
   submit() {
-    if (this.notesid) 
-    {
-      if(this.myform.valid)
-      {
+    if (this.notesid) {
+      if (this.myform.valid) {
         this.service.updateNotes(this.myform.value).subscribe(res => {
           this.messageService.add({ severity: 'info', summary: 'Notes Updated', detail: res.msg });
         })
         this.rut.navigateByUrl("/faculty/faculty-add-notes");
       }
-      else
-      {
+      else {
         this.messageService.add({ severity: 'error', summary: 'Error', detail: 'Invalid credentials' });
       }
-      
+
     }
-    else 
-    {
-      if(this.myform.valid)
-      {
+    else {
+      if (this.myform.valid) {
         this.service.addNotes(this.myform.value).subscribe(res => {
           this.messageService.add({ severity: 'success', summary: 'Notes Added', detail: res.msg });
         })
         this.rut.navigateByUrl("/faculty/faculty-add-notes");
       }
-      else
-      {
+      else {
         this.messageService.add({ severity: 'error', summary: 'Error', detail: 'Invalid credentials' });
       }
-      
+
     }
 
 
   }
 
   delete(value) {
-    this.service.deleteNotes(value).subscribe(res => {
-      console.log("deleted...");
 
-    })
+    this.confirmationService.confirm({
+      message: 'Do you want to delete this record?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.service.deleteNotes(value).subscribe(res => {
+          this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'Record deleted' });
+
+        })
+      },
+      reject: (type) => {
+        switch (type) {
+          case ConfirmEventType.REJECT:
+            this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+            break;
+          case ConfirmEventType.CANCEL:
+            this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'You have cancelled' });
+            break;
+        }
+      }
+    });
+
   }
 
 }
