+++ src/app/student-list-feedback/student-list-feedback.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { ConfirmationService, ConfirmEventType, MessageService } from 'primeng/api';
 import { DataService } from '../data.service';
 
 @Component({
@@ -10,7 +11,7 @@ export class StudentListFeedbackComponent implements OnInit {
 
   feedback:{};
   userid=0;
-  constructor(private service:DataService) { }
+  constructor(private service:DataService, private confirmationService: ConfirmationService, private messageService: MessageService) { }
 
   async ngOnInit()
   {
@@ -23,10 +24,29 @@ export class StudentListFeedbackComponent implements OnInit {
   }
   delete(value)
   {
-    this.service.deletefeedback(value).subscribe(res =>{
-      console.log("deleted",res);
-      
-    })
+    this.confirmationService.confirm({
+      message: 'Do you want to delete this record?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.service.deletefeedback(value).subscribe(res =>{
+          this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'Record deleted' });
+        })
+
+
+      },
+      reject: (type) => {
+        switch (type) {
+          case ConfirmEventType.REJECT:
+            this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+            break;
+          case ConfirmEventType.CANCEL:
+            this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'You have cancelled' });
+            break;
+        }
+      }
+    });
+    
   }
 
 }
